Add files via upload
Todo App
This is a simple Todo application built with React and Redux Toolkit. It allows users to add and remove tasks from a list.

Features
Add tasks: Users can enter a task in the input field and press the "Add Task" button to add it to the list.
Remove tasks: Each task in the list has a "Remove" button that allows users to remove it from the list.
Responsive design: The app is designed to be responsive and works well on desktop and mobile devices.
Redux state management: The app uses Redux Toolkit for state management, making it easy to manage the application's state and data flow.
Technologies Used
React.js
Redux Toolkit
Bootstrap (for styling)
How to Run
Clone the repository to your local machine.
Install dependencies using npm install.
Run the app using npm start.
Open http://localhost:3000 to view it in the browser.